Refactor instrumentation setup functions into separate modules
